Add required userAgent
The requirements of Nominatim state that a User-Agent must be submitted (https://operations.osmfoundation.org/policies/nominatim/). It looks like this was already implemented in the provider but not in the config (`ProviderFactory`). So `Geocoder\Provider\Nominatim` does now require a userAgent as third parameter.

Solves Issue: https://github.com/geocoder-php/BazingaGeocoderBundle/issues/195