Add required userAgent (#198)
* Add required userAgent

The requirements of Nominatim state that a User-Agent must be submitted (https://operations.osmfoundation.org/policies/nominatim/). It looks like this was already implemented in the provider but not in the config (`ProviderFactory`). So `Geocoder\Provider\Nominatim` does now require a userAgent as third parameter.

Solves Issue: https://github.com/geocoder-php/BazingaGeocoderBundle/issues/195

* Added mandatory parameter to test config

* Made test running again by setting a default user_agent

* Made user_agent option as required